Docker image/requirements.txt
torch==1.9.1
torchvision==0.10.1
Pillow==8.3.2